move thrift defs to shared.thrift
