+++ .gen/go/history/historyservicetest/client.go
@@ -71,9 +71,9 @@ func (m *MockClient) EXPECT() *_MockClientRecorder {
 // 	... := client.DescribeHistoryHost(...)
 func (m *MockClient) DescribeHistoryHost(
 	ctx context.Context,
-	_Request *history.DescribeHistoryHostRequest,
+	_Request *shared.DescribeHistoryHostRequest,
 	opts ...yarpc.CallOption,
-) (success *history.DescribeHistoryHostResponse, err error) {
+) (success *shared.DescribeHistoryHostResponse, err error) {
 
 	args := []interface{}{ctx, _Request}
 	for _, o := range opts {
@@ -81,7 +81,7 @@ func (m *MockClient) DescribeHistoryHost(
 	}
 	i := 0
 	ret := m.ctrl.Call(m, "DescribeHistoryHost", args...)
-	success, _ = ret[i].(*history.DescribeHistoryHostResponse)
+	success, _ = ret[i].(*shared.DescribeHistoryHostResponse)
 	i++
 	err, _ = ret[i].(error)
 	return
