+++ idl/github.com/uber/cadence/shared.thrift
@@ -1113,6 +1113,26 @@ struct DescribeTaskListResponse {
   10: optional list<PollerInfo> pollers
 }
 
+//At least one of the parameters needs to be provided
+struct DescribeHistoryHostRequest {
+  10: optional string hostAddress //ip:port
+  20: optional i32 shardIdForHost
+  30: optional WorkflowExecution executionForHost
+}
+
+struct DescribeHistoryHostResponse{
+  10: optional i32 numberOfShards
+  20: optional list<i32> shardIDs
+  30: optional DomainCacheInfo domainCache
+  40: optional string shardControllerStatus
+  50: optional string address
+}
+
+struct DomainCacheInfo{
+  10: optional i64 numOfItemsInCacheByID
+  20: optional i64 numOfItemsInCacheByName
+}
+
 enum TaskListType {
   /*
    * Decision type of tasklist
