+++ service/frontend/adminHandler.go
@@ -117,7 +117,7 @@ func (adh *AdminHandler) DescribeWorkflowExecution(ctx context.Context, request 
 }
 
 // DescribeHistoryHost returns information about the internal states of a history host
-func (adh *AdminHandler) DescribeHistoryHost(ctx context.Context, request *admin.DescribeHistoryHostRequest) (*admin.DescribeHistoryHostResponse, error) {
+func (adh *AdminHandler) DescribeHistoryHost(ctx context.Context, request *gen.DescribeHistoryHostRequest) (*gen.DescribeHistoryHostResponse, error) {
 	if request == nil || (request.ShardIdForHost == nil && request.ExecutionForHost == nil && request.HostAddress == nil) {
 		return nil, adh.error(errRequestNotSet)
 	}
@@ -128,24 +128,8 @@ func (adh *AdminHandler) DescribeHistoryHost(ctx context.Context, request *admin
 		}
 	}
 
-	resp, err := adh.history.DescribeHistoryHost(ctx, &hist.DescribeHistoryHostRequest{
-		HostAddress:      request.HostAddress,
-		ShardIdForHost:   request.ShardIdForHost,
-		ExecutionForHost: request.ExecutionForHost,
-	})
-	if err != nil {
-		return &admin.DescribeHistoryHostResponse{}, err
-	}
-	return &admin.DescribeHistoryHostResponse{
-		NumberOfShards: resp.NumberOfShards,
-		ShardIDs:       resp.ShardIDs,
-		DomainCache: &admin.DomainCache{
-			NumOfItemsInCacheByID:   resp.DomainCache.NumOfItemsInCacheByID,
-			NumOfItemsInCacheByName: resp.DomainCache.NumOfItemsInCacheByName,
-		},
-		ShardControllerStatus: resp.ShardControllerStatus,
-		Address:               resp.Address,
-	}, err
+	resp, err := adh.history.DescribeHistoryHost(ctx, request)
+	return resp, err
 }
 
 func (adh *AdminHandler) error(err error) error {
